added delete restriction for non admin

+++ src/Components/Context/JobProvider.jsx
@@ -6,15 +6,18 @@ import Swal from 'sweetalert2';
 
 export const JobContext = createContext(null)
 const JobProvider = ({ children }) => {
+
+    const [isAdmin, setIsAdmin] = useState(false)
     const [jobs, setJobs] = useState([])
+    const [users, setUsers] = useState([])
     const [prevJobs, setPrevJobs] = useState([])
     const [isLoading, setIsLoading] = useState(true)
     const [isOpen, setIsOpen] = useState(false)
     const [selectedJobForPartialDelivery, setSelectedJobForPartialDelivery] = useState(null);
     const [partialDeliveryQty, setPartialDeliveryQty] = useState(0);
 
-    const baseUrl = "https://delivery-report-yunusco-back-end.vercel.app"
-    // const baseUrl = "http://localhost:8570"
+    // const baseUrl = "https://delivery-report-yunusco-back-end.vercel.app"
+    const baseUrl = "http://localhost:8570"
 
     useEffect(() => {
         axios.get(`${baseUrl}/delivery`)
@@ -33,8 +36,28 @@ const JobProvider = ({ children }) => {
                 setIsLoading(false)
             })
     }, [prevJobs])
+    // get all the users
+    useEffect(() => {
+        axios.get(`${baseUrl}/users`)
+            .then(res => {
+                setIsLoading(true)
+                setUsers(res.data)
+                setIsLoading(false)
+            })
+    }, [users])
 
-
+    // get logged user only
+    const handleAdminSearch = (email) => {
+        axios.get(`${baseUrl}/isAdmin/${email}`)
+            .then(res => {
+                setIsLoading(true)
+                
+                if(res.data.role === "admin"){
+                    setIsAdmin(true)
+                }
+                setIsLoading(false)
+            })
+    }
     const addUser = async (user) => {
         try {
             await axios.post(`${baseUrl}/postUser`, user);
@@ -48,13 +71,7 @@ const JobProvider = ({ children }) => {
             // });
         } catch (error) {
             if (error.response && error.response.status === 400) {
-                Swal.fire({
-                    position: 'top-center',
-                    icon: 'error',
-                    title: 'Email Already Exists',
-                    text: 'user with this email already exists.',
-                    showConfirmButton: "OK"
-                });
+                return
             }
             console.error("Error Adding user:", error);
             // Handle error and show a message to the user
@@ -88,6 +105,7 @@ const JobProvider = ({ children }) => {
 
     const handleDelete = async (job) => {
         // Show a confirmation dialog using Swal.fire
+
         Swal.fire({
             title: 'Are you sure?',
             text: "You won't be able to revert this!",
@@ -98,28 +116,39 @@ const JobProvider = ({ children }) => {
             confirmButtonText: 'Yes, delete it!'
         }).then(async (result) => {
             if (result.isConfirmed) {
-                try {
-                    await axios.delete(`${baseUrl}/deleteJob/${job._id}`);
-                    setJobs(prevJobs => prevJobs.filter(j => j._id !== job._id));
+                if(isAdmin){
 
-                    // Display SweetAlert success alert
-                    Swal.fire({
-                        icon: 'success',
-                        title: 'Job Deleted',
-                        text: 'The job has been successfully Deleted from On Proccessing delivery.',
-                    });
-                } catch (error) {
-                    if (error) {
+                    try {
+                        await axios.delete(`${baseUrl}/deleteJob/${job._id}`);
+                        setJobs(prevJobs => prevJobs.filter(j => j._id !== job._id));
+    
+                        // Display SweetAlert success alert
                         Swal.fire({
-                            position: 'top-center',
-                            icon: 'error',
-                            title: `${error.message}`,
-                            text: 'An Error Occured during this Oparation',
-                            showConfirmButton: "OK"
+                            icon: 'success',
+                            title: 'Job Deleted',
+                            text: 'The job has been successfully Deleted from On Proccessing delivery.',
                         });
+                    } catch (error) {
+                        if (error) {
+                            Swal.fire({
+                                position: 'top-center',
+                                icon: 'error',
+                                title: `${error.message}`,
+                                text: 'An Error Occured during this Oparation',
+                                showConfirmButton: "OK"
+                            });
+                        }
+                        console.error("Error delivering job:", error);
+                        // Handle error and show a message to the user
                     }
-                    console.error("Error delivering job:", error);
-                    // Handle error and show a message to the user
+                } else{
+                    Swal.fire({
+                        position: 'top-center',
+                        icon: 'error',
+                        title: "Unauthorize Oparetion",
+                        text: "Look Like You Don't have the permission to Delete Job",
+                        showConfirmButton: "OK"
+                    });
                 }
             }
         });
@@ -151,7 +180,7 @@ const JobProvider = ({ children }) => {
                         Swal.fire({
                             position: 'top-center',
                             icon: 'error',
-                            title:`404
+                            title: `404
                             NOT FOUND`,
                             text: `${error.message}`,
                             showConfirmButton: "OK"
@@ -180,6 +209,7 @@ const JobProvider = ({ children }) => {
         partialDeliveryQty,
         setPartialDeliveryQty,
         addUser,
+        handleAdminSearch,
         isLoading
     }
     return (
